refactor(skills): centralize internal write paths
