+++ apps/android/app/src/main/java/ai/openclaw/app/gateway/GatewayProtocol.kt
@@ -289,11 +289,6 @@ enum class GatewayMethod(
   SkillsCuratorPin("skills.curator.pin"),
   SkillsCuratorUnpin("skills.curator.unpin"),
   SkillsCuratorRestore("skills.curator.restore"),
-  SkillsWriteValidate("skills.write.validate"),
-  SkillsWritePropose("skills.write.propose"),
-  SkillsWriteApplyProposal("skills.write.applyProposal"),
-  SkillsWriteDirect("skills.write.direct"),
-  SkillsWriteRefreshSnapshot("skills.write.refreshSnapshot"),
   SkillsProposalsList("skills.proposals.list"),
   SkillsProposalsInspect("skills.proposals.inspect"),
   SkillsProposalsHistoryStatus("skills.proposals.historyStatus"),
