+++ packages/gateway-protocol/src/index.ts
@@ -536,16 +536,6 @@ import {
   SkillsStatusParamsSchema,
   SkillsUploadBeginParamsSchema,
   SkillsUploadChunkParamsSchema,
-  SkillsWriteDirectParamsSchema,
-  SkillsWriteDirectResultSchema,
-  SkillsWriteApplyProposalParamsSchema,
-  SkillsWriteApplyProposalResultSchema,
-  SkillsWriteProposeParamsSchema,
-  SkillsWriteProposeResultSchema,
-  SkillsWriteRefreshSnapshotParamsSchema,
-  SkillsWriteRefreshSnapshotResultSchema,
-  SkillsWriteValidateParamsSchema,
-  SkillsWriteValidateResultSchema,
   SkillsUploadCommitParamsSchema,
   SkillsUpdateParamsSchema,
   ToolsCatalogParamsSchema,
@@ -900,15 +890,6 @@ export const validateSkillsProposalRequestRevisionParams = lazyCompile(
   SkillsProposalRequestRevisionParamsSchema,
 );
 export const validateSkillsProposalActionParams = lazyCompile(SkillsProposalActionParamsSchema);
-export const validateSkillsWriteValidateParams = lazyCompile(SkillsWriteValidateParamsSchema);
-export const validateSkillsWriteProposeParams = lazyCompile(SkillsWriteProposeParamsSchema);
-export const validateSkillsWriteApplyProposalParams = lazyCompile(
-  SkillsWriteApplyProposalParamsSchema,
-);
-export const validateSkillsWriteDirectParams = lazyCompile(SkillsWriteDirectParamsSchema);
-export const validateSkillsWriteRefreshSnapshotParams = lazyCompile(
-  SkillsWriteRefreshSnapshotParamsSchema,
-);
 export const validateSkillsSecurityVerdictsParams = lazyCompile(SkillsSecurityVerdictsParamsSchema);
 export const validateSkillsSkillCardParams = lazyCompile(SkillsSkillCardParamsSchema);
 export const validateCronListParams = lazyCompile(CronListParamsSchema);
@@ -1376,16 +1357,6 @@ export {
   SkillsProposalActionParamsSchema,
   SkillsProposalApplyResultSchema,
   SkillsProposalRecordResultSchema,
-  SkillsWriteDirectParamsSchema,
-  SkillsWriteDirectResultSchema,
-  SkillsWriteApplyProposalParamsSchema,
-  SkillsWriteApplyProposalResultSchema,
-  SkillsWriteProposeParamsSchema,
-  SkillsWriteProposeResultSchema,
-  SkillsWriteRefreshSnapshotParamsSchema,
-  SkillsWriteRefreshSnapshotResultSchema,
-  SkillsWriteValidateParamsSchema,
-  SkillsWriteValidateResultSchema,
   SkillsSecurityVerdictsParamsSchema,
   SkillsSecurityVerdictsResultSchema,
   SkillsSkillCardParamsSchema,
@@ -1757,16 +1728,6 @@ export type {
   SkillsProposalActionParams,
   SkillsProposalApplyResult,
   SkillsProposalRecordResult,
-  SkillsWriteDirectParams,
-  SkillsWriteDirectResult,
-  SkillsWriteApplyProposalParams,
-  SkillsWriteApplyProposalResult,
-  SkillsWriteProposeParams,
-  SkillsWriteProposeResult,
-  SkillsWriteRefreshSnapshotParams,
-  SkillsWriteRefreshSnapshotResult,
-  SkillsWriteValidateParams,
-  SkillsWriteValidateResult,
   SkillsSecurityVerdictsParams,
   SkillsSecurityVerdictsResult,
   SkillsSkillCardParams,
