+++ src/gateway/methods/core-descriptors.ts
@@ -191,11 +191,6 @@ const CORE_GATEWAY_METHOD_SPECS: readonly CoreGatewayMethodSpec[] = [
   { name: "skills.curator.pin", scope: "operator.admin", since: "<=2026.7" },
   { name: "skills.curator.unpin", scope: "operator.admin", since: "<=2026.7" },
   { name: "skills.curator.restore", scope: "operator.admin", since: "<=2026.7" },
-  { name: "skills.write.validate", scope: "operator.read", since: "2026.7" },
-  { name: "skills.write.propose", scope: "operator.admin", since: "2026.7" },
-  { name: "skills.write.applyProposal", scope: "operator.admin", since: "2026.7" },
-  { name: "skills.write.direct", scope: "operator.admin", since: "2026.7" },
-  { name: "skills.write.refreshSnapshot", scope: "operator.admin", since: "2026.7" },
   { name: "skills.proposals.list", scope: "operator.read", since: "<=2026.7" },
   { name: "skills.proposals.inspect", scope: "operator.read", since: "<=2026.7" },
   { name: "skills.proposals.historyStatus", scope: "operator.read", since: "<=2026.7" },
