+++ src/gateway/server-methods/skills-method-names.ts
@@ -14,11 +14,6 @@ export const SKILLS_GATEWAY_METHOD_NAMES = [
   "skills.curator.pin",
   "skills.curator.unpin",
   "skills.curator.restore",
-  "skills.write.validate",
-  "skills.write.propose",
-  "skills.write.applyProposal",
-  "skills.write.direct",
-  "skills.write.refreshSnapshot",
   "skills.proposals.list",
   "skills.proposals.inspect",
   "skills.proposals.historyStatus",
