+++ src/gateway/server-methods/skills-proposal-history.ts
@@ -4,12 +4,12 @@ import {
 } from "../../../packages/gateway-protocol/src/schema/skill-history.js";
 import { getSkillHistoryScanStatus } from "../../skills/workshop/history-scan-state.js";
 import { runSkillHistoryScan } from "../../skills/workshop/history-scan.js";
-import { runSkillsWorkspaceHandler } from "./skills-workspace-handler.js";
+import { runSkillsProposalWorkspaceHandler } from "./skills-workspace-handler.js";
 import type { GatewayRequestHandlers } from "./types.js";
 
 export const skillProposalHistoryHandlers: GatewayRequestHandlers = {
   "skills.proposals.historyStatus": async ({ params, respond, context }) => {
-    await runSkillsWorkspaceHandler({
+    await runSkillsProposalWorkspaceHandler({
       method: "skills.proposals.historyStatus",
       rawParams: params,
       respond,
@@ -26,7 +26,7 @@ export const skillProposalHistoryHandlers: GatewayRequestHandlers = {
     });
   },
   "skills.proposals.historyScan": async ({ params, respond, context }) => {
-    await runSkillsWorkspaceHandler({
+    await runSkillsProposalWorkspaceHandler({
       method: "skills.proposals.historyScan",
       rawParams: params,
       respond,
