--- src/skills/api/index.ts
@@ -1,12 +1,0 @@
-export { skillsWriteService } from "./service.js";
-export type {
-  SkillsWriteDirectInput,
-  SkillsWriteDirectResult,
-  SkillsWriteDirectRollback,
-  SkillsWriteInstallBundleInput,
-  SkillsWriteInstallBundleResult,
-  SkillsWriteProposalInput,
-  SkillsWriteService,
-  SkillsWriteValidationInput,
-  SkillsWriteValidationResult,
-} from "./types.js";
