+++ src/skills/lifecycle/workspace-skill-write.ts
@@ -23,11 +23,6 @@ type WorkspaceSkillWriteTargetParams = {
 
 type PreviousSupportFile = { path: string; existed: boolean; previousContent?: string };
 
-export type WorkspaceSkillWriteRollback = {
-  previousContent: string | null;
-  supportFiles: PreviousSupportFile[];
-};
-
 export function normalizeWorkspaceSkillSupportPath(input: string): string {
   const trimmed = input.trim();
   if (!trimmed) {
@@ -73,17 +68,14 @@ export function assertWorkspaceSkillSupportPathSetIsFileOnly(paths: readonly str
   }
 }
 
-export async function readWorkspaceSkillFile(
-  filePath: string,
-  options: { maxBytes?: number } = {},
-): Promise<string | null> {
+export async function readWorkspaceSkillFile(filePath: string): Promise<string | null> {
   if (!(await pathExists(filePath))) {
     return null;
   }
   const skillRoot = await root(path.dirname(filePath));
   const read = await skillRoot.read(path.basename(filePath), {
     hardlinks: "reject",
-    maxBytes: options.maxBytes ?? 1024 * 1024,
+    maxBytes: 1024 * 1024,
     symlinks: "reject",
   });
   return read.buffer.toString("utf8");
@@ -119,18 +111,16 @@ export async function writeWorkspaceSkill(params: {
   content: string;
   supportFiles?: readonly WorkspaceSkillSupportFileWrite[];
   mode: "create" | "update";
-  maxPreviousSkillFileBytes?: number;
   symlinkPolicy: WorkspaceSkillSymlinkWritePolicy;
-}): Promise<WorkspaceSkillWriteRollback> {
+}): Promise<void> {
   assertInsideWorkspace(params.workspaceDir, params.skillDir, "skill directory");
   const supportFiles = normalizeSupportFiles(params.supportFiles ?? []);
-  const previous = await prepareWorkspaceSkillWrite({
+  const previousSupportFiles = await prepareWorkspaceSkillWrite({
     mode: params.mode,
     workspaceDir: params.workspaceDir,
     skillDir: params.skillDir,
     skillFile: params.skillFile,
     supportFiles,
-    maxPreviousSkillFileBytes: params.maxPreviousSkillFileBytes,
     symlinkPolicy: params.symlinkPolicy,
   });
   const writtenSupportPaths: string[] = [];
@@ -158,12 +148,11 @@ export async function writeWorkspaceSkill(params: {
       workspaceDir: params.workspaceDir,
       skillDir: params.skillDir,
       writtenSupportPaths,
-      previousSupportFiles: previous.supportFiles,
+      previousSupportFiles,
       symlinkPolicy: params.symlinkPolicy,
     });
     throw error;
   }
-  return previous;
 }
 
 function normalizeSupportFiles(
@@ -183,17 +172,14 @@ async function prepareWorkspaceSkillWrite(params: {
   skillDir: string;
   skillFile: string;
   supportFiles: readonly WorkspaceSkillSupportFileWrite[];
-  maxPreviousSkillFileBytes?: number;
   symlinkPolicy: WorkspaceSkillSymlinkWritePolicy;
-}): Promise<WorkspaceSkillWriteRollback> {
+}): Promise<PreviousSupportFile[]> {
   await resolveWorkspaceSkillWriteTarget({
     workspaceDir: params.workspaceDir,
     filePath: params.skillFile,
     symlinkPolicy: params.symlinkPolicy,
   });
-  const previousContent = await readWorkspaceSkillFile(params.skillFile, {
-    maxBytes: params.maxPreviousSkillFileBytes,
-  });
+  const previousContent = await readWorkspaceSkillFile(params.skillFile);
   if (params.mode === "create" && previousContent !== null) {
     throw new Error(`Target skill already exists: ${params.skillFile}`);
   }
@@ -209,11 +195,7 @@ async function prepareWorkspaceSkillWrite(params: {
       filePath,
       symlinkPolicy: params.symlinkPolicy,
     });
-    if (params.mode === "create") {
-      // Create writes use overwrite=false, so any pre-existing target aborts before this
-      // metadata is returned; a successful create can safely remove these files on rollback.
-      previousSupportFiles.push({ path: file.path, existed: false });
-    } else {
+    if (params.mode === "update") {
       const previousSupportContent = await readWorkspaceSupportFile({
         skillDir: params.skillDir,
         relativePath: file.path,
@@ -225,7 +207,7 @@ async function prepareWorkspaceSkillWrite(params: {
       );
     }
   }
-  return { previousContent, supportFiles: previousSupportFiles };
+  return previousSupportFiles;
 }
 
 async function writeWorkspaceFile(params: {
